% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchfork_reviews.R
\docType{data}
\name{pitchfork_reviews}
\alias{pitchfork_reviews}
\title{Pitchfork Album Reviews}
\format{A data frame with 979 rows and 5 variables:
\describe{
  \item{artist}{artist's name}
  \item{score}{score given by pitchfork, from a scale of 0 to 10}
  \item{album}{album name}
  \item{bnm}{dummy variable, 1 if the album received Best New Music or Best New Reissue, 0 otherwise}
  \item{pub_date}{review publication date}
  ...
}}
\source{
\url{https://pitchfork.com/}
}
\usage{
pitchfork_reviews
}
\description{
A dataset containing the score given by Pitchfork to albums released by artists having reached Spotify's top 200 tracks (U.S.).
}
\keyword{datasets}
